adding onMessageDeleted to message receiver
