+++ application/src/main/java/org/togetherjava/tjbot/commands/code/FormatCodeCommand.java
@@ -1,4 +1,4 @@
-package org.togetherjava.tjbot.commands.formatter;
+package org.togetherjava.tjbot.commands.code;
 
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.entities.MessageEmbed;
@@ -11,18 +11,10 @@ import org.togetherjava.tjbot.commands.MessageContextCommand;
 import org.togetherjava.tjbot.formatter.Formatter;
 import org.togetherjava.tjbot.formatter.tokenizer.Lexer;
 
-import java.awt.Color;
 import java.util.Optional;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 // TODO Doc
 public final class FormatCodeCommand extends BotCommandAdapter implements MessageContextCommand {
-    private static final Color AMBIENT_COLOR = Color.decode("#FDFD96");
-    // TODO doc, lol
-    private static final Pattern CODE_BLOCK_EXTRACTOR_PATTERN =
-            Pattern.compile("```(?:java)?\\s*([\\w\\W]+)```|``?([\\w\\W]+)``?");
-
     private final Lexer lexer;
     private final Formatter formatter;
 
@@ -51,14 +43,6 @@ public final class FormatCodeCommand extends BotCommandAdapter implements Messag
             .build();
 
         event.replyEmbeds(response).queue();
-    }
-
-    private static Optional<String> extractCode(CharSequence fullMessage) {
-        Matcher codeBlockMatcher = CODE_BLOCK_EXTRACTOR_PATTERN.matcher(fullMessage);
-        if (!codeBlockMatcher.find()) {
-            return Optional.empty();
-        }
-        return Optional.of(codeBlockMatcher.group(1));
     }
 
     private String formatCode(String code) {
