+++ application/src/main/java/org/togetherjava/tjbot/commands/code/package-info.java
@@ -1,0 +1,10 @@
+/**
+ * This package offers all the functionality for working with messages containing code.
+ */
+@MethodsReturnNonnullByDefault
+@ParametersAreNonnullByDefault
+package org.togetherjava.tjbot.commands.code;
+
+import org.togetherjava.tjbot.annotations.MethodsReturnNonnullByDefault;
+
+import javax.annotation.ParametersAreNonnullByDefault;
