--- application/src/main/java/org/togetherjava/tjbot/commands/formatter/package-info.java
@@ -1,11 +1,0 @@
-/**
- * This package offers all the functionality for code formatting. The core class is
- * {@link org.togetherjava.tjbot.commands.formatter.FormatCodeCommand}.
- */
-@MethodsReturnNonnullByDefault
-@ParametersAreNonnullByDefault
-package org.togetherjava.tjbot.commands.formatter;
-
-import org.togetherjava.tjbot.annotations.MethodsReturnNonnullByDefault;
-
-import javax.annotation.ParametersAreNonnullByDefault;
